requirements.txt
feedparser>=6.0.10
requests>=2.32.3
beautifulsoup4>=4.12.3
google-generativeai>=0.6.0
python-dotenv>=1.0.1
